Extract tests for className to own file
